feat: add created at sort in rag
